plugin: Add dynamic support for CN version of FFXIV (#825)
